Fix a bug where constant elision fires erroneously.
While specializing a block B to run in its parent A's scope, if we
discover a reference to a symbol which is constant in B's scope, the
current optimizer elides the reference. This is invalid, because the
optimized code will be referenced by A, in an instruction which creates
siblings of B. But not every sibling of B need have the same constant
value for the symbol, so baking the value into the bytecode does not work.