Process escaped symbols in the lexer, not the parser.
This prevents the formation of trivial parse trees from the
reduction

  Symbol -> symbol

where the outer Symbol node is unescaped and the inner terminal
node is not. This, in turn, means that users of these parse trees
can access the unescaped version of the symbol, which allows macro
invokers to pass in a symbol which is then unescaped and used as a
non-symbol token in the expansion of the macro.