Track specialization of procedures, not objects.
Procedures are not 1:1 with objects: many objects can share the
same procedure, and executing in the scope of an object does not
even necessarily mean that we are executing that object's procedure,
if the object's procedure has been respecialized since we started
executing its old procedure.

Since procedures are really the things that get specialized, it makes
more sense to track the specialization of procedures, not objects.
This fixes a bug where we would modify in place the current procedure
as long as the current scope object was specialized, not taking into
account that the current procedure might not be the procedure of the
current scope object.