Correctly attribute ownership of captures in the optimizer.
We were adding captured objects to the optimizer's captures stack
in the wrong order with respect to the chain of ownership, so the
owner object associated with each capture was incorrect. This led
to the wrong captures being optimized away because their owner was
an ancestor of the specialization scope.