Only lex tokens which are expected by the parser.
This addresses # outstanding design problems:
1. Tokens in DSLs should not affect parsing outside the DSL (e.g.
`def` should never be lexed as a keywords outside of a python DSL).
2. String literals are now possible: since tokens are now scoped,
there's nothing wrong with defining 1-character "char" tokens
expected only between "". This will not prevent one-character symbol
names (e.g. x, y, i, j, k) from being lexed as symbols outside of "".
3. The lexer can now produce a symbol if other tokens match but cause
a parse error, which is a quality-of-life change I have been
considering for some time as eliminates some uses of ``. Better, this
change accomplishes this in a much more elegant way than hard-coding
a fallback case into the lexer.
4. We no longer have the weird behavior where new tokens are lexable
immediately but macros are not parsable until we rebuild the automaton
(after a low-precedence reduction). Since lexing is now constrained by
the state machine (not the grammar!) new tokens will not be lexed
until the reductions that use them are ready to be applied.

This is also a first step towards a solution for the biggest outstanding
design problem: how to stop special character symbols from being
"de-symbolified" in the definition of a macro; e.g., what if we want to
use `{` as a symbol (not as the token LBRACE) in the definition of a
macro? This new feature is a step towards a quote operator ' which is
defined by
    7 -> ' <symbol>
Since <symbol> is the only token expected after ', it forces whatever
follows to be parsed as a symbol. This is like adding ``, but because
' and the symbol following it are separate tokens, we can escape a
symbol in a macro definition by prepending ' without needing `` at the
usage site (and we can have the prelude macro facilities do this
automatically).

This problem is still not completely solved. Further work that is
required:
1. We can't actually define the macro body of ' yet. It should just
return the symbol which followed ', but using this symbol in a macro
definition will de-symbolify it! To fix this, we need to allow macro
definitions to specify the non-terminal of the returned parse tree,
rather than defaulting it to the start symbol. This way, ' can return
a parse tree whose non-terminal only has one production, which only
takes a symbol, which will then force the symbol to be lexed as a
symbol using this feature. This is a big hammer to add to the language
just to solve this problem (since parsing non-start nonterminals will
require significant changes to the parser). But it will be generally
useful. For example, some DSLs may want to define macros that don't
get parsed as expressions at all, but rather get parsed as some other
non-terminal specifc to that DSL.
2. Currently, symbols in macro definitions are lexed eagerly (using
TokenTrie_GetToken) in a pass that completes before parsing. This
will need to change, so that these symbols can be lexed in a context-
sensitive way just like regular tokens. Essentially, we need to do for
the STREAM_VECTOR ParseTreeStream implementation what this change does
for the STREAM_LEXER implementation. We should be able to reuse most of
the code, though.