Make macros behave more consistently with respect to non-terminals.
This change aims to have the parser respect the non-terminal return
value of parse tree objects in macro expansions. Instead of parsing
as an expression (precedence 1) and then converting to a term
(precedence 7) (which are the most permissive for parsing and using,
respectively), the parse tree returned by a macro handler is now
parsed as the non-terminal symbol returned by the tree, and converted
to the non-terminal of the macro production. This allows the resulting
parse tree to be used anywhere the non-terminal of the macro production
could be used (even if it is an ad-hoc non-terminal not part of the normal
expression tree, and thus has no path to the non-terminal `1`). Likewise,
it allows the user to choose the grammar according to which the tree is
parsed, even if the tree conforms to some ad-hoc grammar but not to the
normal expression grammar.

Since the non-terminal of a parse tree is now used in reparsing, and since
macro expansions have the parse tree returned by the macro but with the
non-terminal of the macro production, this breaks reparsing: if we try to
reparse a parse tree from a macro expansion when it is used as a nested
tree in the expansion of another macro, we may fail because it's non-terminal
is now different from the non-terminal by which it was originally parsed. To
fix this, we simply remove reparsing of nested parse trees. Instead, we just
treat them as representing primitive expression ASTs and error if they don't.
This is a much-needed improvement anyways, because it was horribly inefficient
to continuously reparse the same expressions over and over again. We can refactor
in the future to convert everything to an expression at the end (rather than calling
ParseTree_ReconstructExpr repeatedly, as we do now) which will allow us to traverse
each parse tree exactly once.

Removing reparsing except for the top-level parse tree returned by the macro
also greatly simplifies the semantics of the parser, and in the formal semantic
model it allows us to equate parse tree S-expressions with expr S-expressions.
This is a step towards removing reparsing altogether, so that macros will simply
replace one parse tree with another. Reparsing as a convenience feature can be
added in the standard library by implementing part or all of a bl:mp parser in
bl:mp.

This change also removes the strange implicit behavior where parse tree objects
would elide reductions with only a single sub-tree. This should make it much more
intuitive to deal with manually-constructed parse trees, now that parse trees can
be manually constructed.

Finally, since we no longer parse all macro expansions as expressions (1), macros
can return bare symbols which will be interpreted as entire parse trees on their
own, which allows us (with the last commit) to create a macro which always parses
its argument as a symbol, and then returns that symbol without interpreting it,
even if it coincides with a non-symbol terminal.

Future work that is required to be fully consistent about respecting non-terminals:
the non-terminal of the parse tree used as a macro production should determine the
non-terminal production, which will allow us to eliminate the first sub-expression
of macro definition syntax, which is currently used as the non-terminal symbol.