Optimize away sends of unused messages.
If a message is sent to a block that does not use its message, just
evaluate the block's code without putting the message on the result
stack. The computation of the message can be elided entirely if it
is pure, and the block can now be inlined even if the message is not
pure.

This partially addresses #11, by making
	A ; B -> {{.}} A B
a viable translation for semi-colons that (1) carries no run-time
overhead, and (2) can be implemented without a built-in sequence
statement.

This work is incomplete, however, because we would like to translate
semicolons like
	A ; B -> {{^}} A B
so that a sequence expression has the value of its last component
statement (instead of the arbitrary value `.`). This does not work,
since {^} does use its argument and thus `{^} B` can still not be
inlined if `B` is not a pure value. Further work is needed to fix
this, such as recognizing pure blocks that return their message and
replacing them with the message, or inlining effectful messages into
affine blocks which do not perform side-effects before evaluating their
message.