Inline effectful messages into affine procedures.
This greatly expands the scope of inlining. In particular, it allows
us to always inline when the receiver is {^}, whatever the message
may be, so that `{{^}} A B` compiles to the same code as `A; B`, even
when A or B has side-effects. This is sufficent to close #11, since
it is now possible to implement sequence expressions in the language
with no runtime overhead.

Further work that could improve this feature: currently, we do not
check if the procedure is affine before attempting to inline it. We
inline optimistically and fail if we find we need to evaluate the
message more than once. We could improve worst-case performance by
first checking if the procedure _might be_ affine, which is information
that could be provided during the analysis stage.