Refactor macro tests to be parse-tree-protocol-agnostic.
A new abstraction layer is added which defines new syntax for
creating and using parse trees. Outside of this module, test code
does not need to be aware of the specifics of the parse tree protocol.
This will make it much easier to validate the upcoming breaking change
to the protocol, so we won't have to rework all the tests at the same
time as we make a big change.