Change the parse tree protocol according to #14
The new protocol encodes terminals differently: as objects that
send no sub-trees and which return the terminal symbol, rather than
as the terminal symbol itself. As a result, empty non-terminals
(technically not supported but used as a convention in macro
productions, where only the grammar symbol of the parse tree is
relevant) are represented differently as well, as objects with a
single sub-tree, the empty terminal {``}.

In addition, the mechanism for interpeting the production of a macro
definition is changed. Instead of taking just the top-level grammar
symbols of the production (which is all that is relevant to the macro
definition itself) using an ad hoc visitor, we use the normal
ObjectToParseTree conversion, which deeply evaluates its argument, and
_then_ take only the top-level grammar symbols from the evaluated parse
tree. This is an important semantic distinction when the production has
side effects.

These two changes represent the first two steps of the roadmap to
completing #14.