Smoother handling of multi-line interrupts in REPL.
If the user hits enter twice while in the middle of a multi-line
input, they simply get a blank line, rather than cancelling the
input.

To cancel a multi-line input, the user can now press Ctrl+C,
because the readline interrupt handler no longer just starts the
current line over. Intead, it returns an "interrupted" sentinel,
which allows the caller (Readline_ReadExpr) to start the entire
expression over, instead of just that line.