Optimizer bug fix
Emit(move = true) was assuming, incorrectly, that the offset of
the emitted instruction would be just past the previous end of the
procedure. This is not the case if the expression is not moved due
to already being the last instruction in the procedure. Fixed by
having Bytecode_MoveToEnd return the new offset.

Also, the logic in Bytecode_MoveToEnd for detecting when an
instruction is already last in its procedure was incorrect, due to
a fencepost error. This has been fixed.