Use ! to evaluate std prelude at import time.
This ensures that the std prelude is fully evaluated when any part
of it is used in application code, which means we don't need to worry
about whether std has been evaluated in the expansion of `syntax`,
and can do away with __std_imported.

This is step 5 of the roadmap towards #14.