Bug fix in bang reductions.
Bang expressions produce an object terminal parse tree, whose
structure does not match that of the production which created it
(which is not a terminal at all). This affects the logic which
removes consumed symbols from the set of precedences preventing us
from updating the parser state. Previously, this logic looked at the
produced parse tree fragment's sub-trees to determine which symbols
to remove. Now, it looks at the production itself.