Do not automatically execute bl:mp programs.
With the addition of ! giving full control over execution order to
the programmer, there is no longer a need for exeuction to be a
built-in step in interpreting a bl:mp program. Now, the semantics of
a bl:mp program are defined as the semantics of parsing the program,
and the interpreter (in non-interactive) mode no longer executes the
parsed expression.

The std prelude is modified such that if it is prepended to a file,
it will automatically execute the contents of that file when it is
parsed. The interpreter is extended with a --preload option to prepend
files in this way, and the std prelude is implicitly prepended if
--core is not passed.

This is the first part of the final step towards #14 (the other part
being the analagous change for the test runner).