Remove reparsing from the core language.
The plan has always been to remove the built-in reparsing behavior
of macro expansions and replace it with a pure-bl:mp parser that
can be invoked by macro handlers as required. The first part of
this is done.

The pure-bl:mp parser will require significantly more effort, and
so an interpreter intrinsic that invokes the C parser is currently
standing in its place. This intrinsic must still be explicitly
invoked by macro handlers, and it implements the same interface that
we expect to implement with the pure-bl:mp parser later on.

This allows us to punt on the opaque implementation details of the
bl:mp parser while still structuring things in a way consistent with
the overall design, so that we can proceed with high-level design
problems that depend on the interface of the parser without getting
bogged down in the details of implementing a complicated parser in
bl:mp.

This commit closes #3. We will handle #2 (implementing a pure-bl:mp
parser for bl:mp) later on, after some more high-level design issues
have been worked out and there is time for implementation nitty-gritty.