Use a purpose-built symbol map for scopes.
Instead of using HashMap, which is designed to be very generic and
thus cannot be optimized for a particular type of key, we use a new
type, SymbolMap, which is an efficient bit trie representation of a
sparse set of integers (taking advantage of the fact that symbols
are uniquely identified by their address and can thus easily be
mapped to integers).

This yields a modest few percent performance improvement, but more
importantly there is now room for further optimization. For example,
iteration is very slow in the current implementation, and accounts
for almost 10% of the total time on some benchmarks. In addition,
emplacement is much faster than it is for HashTable, because there
is no expensive reallocation and rehashing cost for growing the set.
Therefore, this may pave the way for optimizing symbol lookups by
caching a pointer to the parent scope which owns a symbol in the
child scope where that symbol is referenced.