Idealized fast iteration for SymbolMap.
Previously, iterating over a SymbolMap required a linear search
for populated sub-trees at each non-empty level. We could skip
empty sub-trees, but a sub-tree that had at least one symbol in
it would be searched linearly for non-empty sub-subtrees. This
was rather inefficient.

With this change, each entry now contains the offset to the next
populated entry at the same level, or 0, which means we only visit
populated sub-trees during the traversal.

The offset only needs to be 8 bits, but SymbolNodeEntry was already
2 words in size, so due to alignment restrictions, the new field
increases the space requirements for a SymbolNode by 50%, which is
unacceptable. This problem can be worked around, though, by reducing
the tag from 8 bytes to 4 bytes, and making it an index into a global
array of symbols, instead of a pointer to a symbol. Since this array
is not yet implemented, this change represents idealized performance
of the algorithm, with the extra space usage.