Add a SymbolMap_Update interface.
The main use of SymbolMap, in object.c, never calls
SymbolMap_AbortEmplace, so having a simpler interface that always
does the update makes sense. For now, SymbolMap_Update calls
SymbolMap_Emplace, but as things get more optimized, we may be able
to make a more efficient implementation that doesn't have to prepare
for a possible revert.