Cache symbols in child scopes that are owned by parent scopes.
This slightly improves the best case performance for symbol lookups.
Upon investigation, further tuning like only caching the symbol in
the child scope where it is directly accessed (as opposed to all
scopes between the child and the owner) or only caching the symbol
when it is owned by a scope a minimum distance away, do not further
improve the performance.

Closes #22